+++ src/config/generic_development_dependency.go
@@ -7,33 +7,20 @@ import (
 	"github.com/Originate/exosphere/src/types"
 )
 
-type genericDependency struct {
-	config    types.DependencyConfig
+type genericDevelopmentDependency struct {
+	config    types.DevelopmentDependencyConfig
 	appConfig types.AppConfig
 	appDir    string
 	homeDir   string
 }
 
 // GetContainerName returns the container name
-func (g *genericDependency) GetContainerName() string {
+func (g *genericDevelopmentDependency) GetContainerName() string {
 	return g.config.Name + g.config.Version
 }
 
-//GetDeploymentConfig returns configuration needed in deployment
-func (g *genericDependency) GetDeploymentConfig() (map[string]string, error) {
-	config := map[string]string{
-		"version": g.config.Version,
-	}
-	return config, nil
-}
-
-// GetDeploymentServiceEnvVariables returns configuration needed for each service in deployment
-func (g *genericDependency) GetDeploymentServiceEnvVariables() map[string]string {
-	return map[string]string{}
-}
-
 // GetDockerConfig returns docker configuration and an error if any
-func (g *genericDependency) GetDockerConfig() (types.DockerConfig, error) {
+func (g *genericDevelopmentDependency) GetDockerConfig() (types.DockerConfig, error) {
 	renderedVolumes, err := tools.GetRenderedVolumes(g.config.Config.Volumes, g.appConfig.Name, g.config.Name, g.homeDir)
 	if err != nil {
 		return types.DockerConfig{}, err
@@ -48,12 +35,12 @@ func (g *genericDependency) GetDockerConfig() (types.DockerConfig, error) {
 }
 
 // GetOnlineText returns the online text
-func (g *genericDependency) GetOnlineText() string {
+func (g *genericDevelopmentDependency) GetOnlineText() string {
 	return g.config.Config.OnlineText
 }
 
 // GetServiceEnvVariables returns the environment variables that need to
 // be passed to services that use it
-func (g *genericDependency) GetServiceEnvVariables() map[string]string {
+func (g *genericDevelopmentDependency) GetServiceEnvVariables() map[string]string {
 	return g.config.Config.ServiceEnvironment
 }