+++ src/types/production_dependency_config.go
@@ -1,0 +1,9 @@
+package types
+
+// ProductionDependencyConfig represents a production dependency
+type ProductionDependencyConfig struct {
+	Config  ProductionDependencyConfigOptions `yaml:",omitempty"`
+	Silent  bool                              `yaml:",omitempty"`
+	Name    string
+	Version string
+}